+++ components/sharing-dialog/types/index.d.ts
@@ -31,6 +31,7 @@ type SharingType =
     | 'dataElementGroup'
     | 'dataElementGroupSet'
     | 'dataSet'
+    | 'dataStore'
     | 'document'
     | 'eventChart'
     | 'eventFilter'
@@ -93,6 +94,8 @@ export interface SharingDialogProps {
      */
     initialSharingSettings?: SharingDialogInitialSharingSettings
     dataSharing?: boolean
+    cascadeDashboardSharing?: boolean
+    preventUsersFromRemovingMetadataWriteAccess?: boolean
     onClose?: ModalOnCloseEventHandler
     onError?: (error: any) => void
     onSave?: () => void